+++ src/main/kotlin/app/revanced/patcher/patch/annotations/PatchAnnotation.kt
@@ -5,20 +5,23 @@ import app.revanced.patcher.patch.Patch
 import kotlin.reflect.KClass
 
 /**
- * Annotation to mark a Class as a patch.
+ * Annotation to mark a class as a patch.
  * @param include If false, the patch should be treated as optional by default.
  */
 @Target(AnnotationTarget.CLASS)
-@Retention(AnnotationRetention.RUNTIME)
-@MustBeDocumented
 annotation class Patch(val include: Boolean = true)
 
 /**
  * Annotation for dependencies of [Patch]es.
  */
 @Target(AnnotationTarget.CLASS)
-@Retention(AnnotationRetention.RUNTIME)
-@MustBeDocumented
 annotation class DependsOn(
     val dependencies: Array<KClass<out Patch<Context>>> = []
 )
+
+
+/**
+ * Annotation to mark [Patch]es which depend on integrations.
+ */
+@Target(AnnotationTarget.CLASS)
+annotation class RequiresIntegrations // required because integrations are decoupled from patches